7term/web/lab3/requirements.txt
python-telegram-bot==21.5
requests==2.32.3